Facebook login added, bug fixing needed
Logging in using facebook adds your Facebook email as GameForest
username, takes your First Name and Last Name on Facebook, sets
"debugger" as password. Bug with FB.logout, bug with page-reload on
log-in, but the session is maintained. Once you log in using FaceBook
you also have a GameForest account associated with it.