REST command OK, unsure how to fix API, read pls
The REST commands are OK, but the API command for retrieving a stat (to
get information like stat id, stat value, etc.) is bugged because of the
ASYNCHRONOUS call. The return value is being returned as undefined
before the ASYNC call finishes executing. Needs to be fixed but not sure
how to fix.